Test Fixes for MavenProfileCfgHistoryNone profile

+++ modules/activiti-engine/src/test/java/org/activiti/engine/test/api/runtime/RuntimeServiceTest.java
@@ -1022,33 +1022,34 @@ public class RuntimeServiceTest extends PluggableActivitiTestCase {
     @Deployment(resources={
     	"org/activiti/engine/test/api/oneTaskProcess.bpmn20.xml"})
     public void testHistoricVariableRemovedWhenRuntimeVariableIsRemoved() {
-    	 Map<String, Object> vars = new HashMap<String, Object>();
-       vars.put("var1", "Hello");
-       vars.put("var2", "World");
-       vars.put("var3", "!");
-       ProcessInstance processInstance = runtimeService.startProcessInstanceByKey("oneTaskProcess", vars);
-       
-       // Verify runtime
-       assertEquals(3, runtimeService.getVariables(processInstance.getId()).size());
-       assertEquals(3, runtimeService.getVariables(processInstance.getId(), Arrays.asList("var1", "var2", "var3")).size());
-       assertNotNull(runtimeService.getVariable(processInstance.getId(), "var2"));
-       
-       // Verify history
-       assertEquals(3, historyService.createHistoricVariableInstanceQuery().list().size());
-       assertNotNull(historyService.createHistoricVariableInstanceQuery().processInstanceId(processInstance.getId()).variableName("var2").singleResult());
-       
-       // Remove one variable
-       runtimeService.removeVariable(processInstance.getId(), "var2");
-       
-       // Verify runtime
-       assertEquals(2, runtimeService.getVariables(processInstance.getId()).size());
-       assertEquals(2, runtimeService.getVariables(processInstance.getId(), Arrays.asList("var1", "var2", "var3")).size());
-       assertNull(runtimeService.getVariable(processInstance.getId(), "var2"));
-       
-       // Verify history
-       assertEquals(2, historyService.createHistoricVariableInstanceQuery().list().size());
-       assertNull(historyService.createHistoricVariableInstanceQuery().processInstanceId(processInstance.getId()).variableName("var2").singleResult());
-       
+	     if(processEngineConfiguration.getHistoryLevel().isAtLeast(HistoryLevel.AUDIT)) {
+	    	 Map<String, Object> vars = new HashMap<String, Object>();
+	       vars.put("var1", "Hello");
+	       vars.put("var2", "World");
+	       vars.put("var3", "!");
+	       ProcessInstance processInstance = runtimeService.startProcessInstanceByKey("oneTaskProcess", vars);
+	       
+	       // Verify runtime
+	       assertEquals(3, runtimeService.getVariables(processInstance.getId()).size());
+	       assertEquals(3, runtimeService.getVariables(processInstance.getId(), Arrays.asList("var1", "var2", "var3")).size());
+	       assertNotNull(runtimeService.getVariable(processInstance.getId(), "var2"));
+	       
+	       // Verify history
+	       assertEquals(3, historyService.createHistoricVariableInstanceQuery().list().size());
+	       assertNotNull(historyService.createHistoricVariableInstanceQuery().processInstanceId(processInstance.getId()).variableName("var2").singleResult());
+	       
+	       // Remove one variable
+	       runtimeService.removeVariable(processInstance.getId(), "var2");
+	       
+	       // Verify runtime
+	       assertEquals(2, runtimeService.getVariables(processInstance.getId()).size());
+	       assertEquals(2, runtimeService.getVariables(processInstance.getId(), Arrays.asList("var1", "var2", "var3")).size());
+	       assertNull(runtimeService.getVariable(processInstance.getId(), "var2"));
+	       
+	       // Verify history
+	       assertEquals(2, historyService.createHistoricVariableInstanceQuery().list().size());
+	       assertNull(historyService.createHistoricVariableInstanceQuery().processInstanceId(processInstance.getId()).variableName("var2").singleResult());
+	    }
     }
     
 }
